feat: vm init hart, pass
w_satp would stuck when given wrong page table or PTE.
In the previous commits, due to "some reason" that i cannot figure it out
before, the kernel data section of vm were start from (etext_addr + PG_SIZE),
which was wrong. After fix it to etext_addr, the w_satp finally
accepted the page table.

The "some reason": actually the "etext" and "trampoline" in vm.rs should be
surround by extern to access the foreign global defined in asm, if not
then the physical address will be overlaped.